refactor: replace lp_manage_position with simplified version
Replace lp_manage_position.py with the simpler rebalancing version:
- Removes target price trigger logic
- Auto-rebalances when price moves out of position bounds
- Switches to single-sided liquidity after first rebalance
- Simpler configuration with just position_width_pct and rebalance_seconds
